requirements.txt
streamlit>=1.28.0
reportlab>=4.0.0
Pillow>=10.0.0
plotly>=5.17.0
qrcode>=7.4.0
requests>=2.31.0
rdflib>=7.0.0
pandas>=2.0.0
Jinja2>=3.1.2
pyyaml>=6.0
retry>=0.9.2
pytest>=7.4.0
notebook>=7.0.0
google-genai>=0.7.0
google-cloud-sql-connector[pg8000]>=1.0.0
psycopg>=3.0.0
